Rename config_params files, move code to keygen

+++ avx2-optimized/config_FG.c
@@ -11,7 +11,6 @@
 #include "fft.c"
 #include "fpr.c"
 
-
 int
 Zf(compute_public)(uint16_t *h,
 	const int8_t *f, const int8_t *g, unsigned logn, uint8_t *tmp)
@@ -1032,11 +1031,11 @@ void sample_FG_sizes(inner_shake256_context *rng, uint8_t *tmp)
 
 	printf("\n");
 	// calculate the average and standard deviation
-	for (int depth = 0; depth <= logn; depth++) {
+	for (size_t depth = 0; depth <= logn; depth++) {
 		double avg = ((double)sum_b[depth]) / nsamples;
 		double stddev = sqrt(((double)sum_bsq[depth]) / nsamples - avg*avg);
 		size_t nr_ints = (int)(avg + 6.0 * stddev + 30) / 31;
-		printf("Depth %d: %.2f (%.2f) --> (%zu ints)\n", depth, avg, stddev, nr_ints);
+		printf("Depth %zu: %.2f (%.2f) --> (%zu ints)\n", depth, avg, stddev, nr_ints);
 	}
 }
